services/webapp/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
pydantic==2.9.2
Pillow==10.4.0
aiokafka==0.10.0
orjson==3.10.7

